Replace counting semaphore with cpp20_counting_semaphore
