+++ libs/pika/synchronization/docs/index.rst
@@ -16,7 +16,7 @@ the C++ standard ones in |pika| threads:
 
 * :cpp:class:`pika::barrier`
 * :cpp:class:`pika::condition_variable`
-* :cpp:class:`pika::lcos::local::counting_semaphore`
+* :cpp:class:`pika::counting_semaphore`
 * :cpp:class:`pika::experimental::event`
 * :cpp:class:`pika::latch`
 * :cpp:class:`pika::mutex`
